Downloads/JAVAPROGRAMS/ExternalPro/requirements.txt
flask==3.0.2
google-generativeai==0.3.2
spotipy==2.23.0
gunicorn==21.2.0 